% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_gam.R
\name{predict.zinbgam}
\alias{predict.zinbgam}
\title{Predict from the fitted model}
\usage{
\method{predict}{zinbgam}(object, newdata, type = c("response", "link"), ...)
}
\arguments{
\item{object}{an object representing a fitted model of class
zinbgam.}

\item{newdata}{data to predict from.}

\item{type}{the type of prediction required.}

\item{...}{additional optional arguments.}
}
\description{
Obtain predictions from a fitted model of class zinbgam.
}
