% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_gam.R
\name{simulate.zipgam}
\alias{simulate.zipgam}
\title{Simulate response from the fitted model}
\usage{
\method{simulate}{zipgam}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{an object representing a fitted model of class zipgam.}

\item{nsim}{number of response vectors to simulate. Defaults to 1.}

\item{seed}{an object specifying if and how the random number
generator should be initialized (‘seeded’).}

\item{...}{additional optional arguments.}
}
\description{
Simulate response from a model of class zipgam.
}

