% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigam.R
\docType{package}
\name{zigam}
\alias{zigam}
\title{zigam: Zero-inflated Generalised Additive Models}
\description{
zigam implements Zero-inflated Generalised Addditive Models (GAMs)
from the mgcv package using the EM Algorithm. This permits Poisson
and Negative Binomial GAMs to fitted using an offset. Confidence
intervals for Possion models can be fitted using a parametric
bootstrap.
}
\details{
To learn more about zigam, start with the vignette:
\code{browseVignettes(package = "zigam")}
}
