% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_gam.R
\name{zipgam}
\alias{zipgam}
\title{Zero-inflated Poisson GAM}
\usage{
zipgam(
  lambda.formula,
  pi.formula,
  data,
  lambda = NULL,
  pi = NULL,
  gamma.pi = 1,
  gamma.lambda = 1,
  min.em = 5,
  max.em = 50,
  tol = 0.01
)
}
\arguments{
\item{lambda.formula}{formula for the count model}

\item{pi.formula}{formula for the binary model}

\item{data}{a data frame or list containing the model
response variable and covariates required by the formula.}

\item{lambda}{initial lambda vector}

\item{pi}{initial pi vector}

\item{gamma.pi}{binary model gamma}

\item{gamma.lambda}{count model gamma}

\item{min.em}{minimum number of EM iterations}

\item{max.em}{maximum number of EM iterations}

\item{tol}{tolerance (default=1.0E-2)}
}
\description{
Fit a zero-inflated Poisson Generalized Additive Model
using the EM Algorithm
}
